Move rank_zero_warn back to distributed
